+++ src/main/kotlin/dev/aaronhowser/mods/aaron/AaronExtraCodecs.kt
@@ -1,9 +1,13 @@
 package dev.aaronhowser.mods.aaron
 
 import com.mojang.serialization.Codec
+import com.mojang.serialization.codecs.RecordCodecBuilder
 import net.minecraft.Util
+import net.minecraft.core.UUIDUtil
 import net.minecraft.network.chat.Component
+import net.minecraft.world.entity.ai.attributes.AttributeModifier
 import org.joml.Vector2d
+import java.util.*
 
 object AaronExtraCodecs {
 
@@ -24,5 +28,29 @@ object AaronExtraCodecs {
 				{ str -> Component.Serializer.fromJson(str) },
 				{ comp -> Component.Serializer.toJson(comp) }
 			)
+
+	val ATTRIBUTE_MODIFIER_CODEC: Codec<AttributeModifier> =
+		RecordCodecBuilder.create { instance ->
+			instance.group(
+				UUIDUtil.CODEC
+					.optionalFieldOf("uuid")
+					.forGetter { modifier -> Optional.of(modifier.id) },
+				Codec.STRING
+					.fieldOf("name")
+					.forGetter(AttributeModifier::getName),
+				Codec.DOUBLE
+					.fieldOf("amount")
+					.forGetter(AttributeModifier::getAmount),
+				Codec.INT
+					.fieldOf("operation")
+					.forGetter { modifier -> modifier.operation.toValue() }
+			).apply(instance) { uuid, name, amount, operationValue ->
+				if (uuid.isPresent) {
+					AttributeModifier(uuid.get(), name, amount, AttributeModifier.Operation.fromValue(operationValue))
+				} else {
+					AttributeModifier(name, amount, AttributeModifier.Operation.fromValue(operationValue))
+				}
+			}
+		}
 
 }